+++ packages/navigation/atlassian-switcher/src/__tests__/unit/component-analytics.tsx
@@ -40,14 +40,7 @@ const DefaultAtlassianSwitcher = (props: any = {}) => {
         href: '/wiki',
       },
     ],
-    fixedLinks: [
-      {
-        key: 'people',
-        label: 'People',
-        Icon: stubIcon,
-        href: '/people',
-      },
-    ],
+    fixedLinks: [],
     adminLinks: [
       {
         key: 'discoverMore',
@@ -125,7 +118,7 @@ describe('Atlassian Switcher - Component Analytics', () => {
         suggestedProducts: ['confluence.ondemand'],
         licensedProducts: ['jira'],
         adminLinks: ['discoverMore'],
-        fixedLinks: ['people'],
+        fixedLinks: [],
         numberOfSites: 2,
       },
     });
@@ -172,7 +165,7 @@ describe('Atlassian Switcher - Component Analytics', () => {
         suggestedProducts: ['confluence.ondemand'],
         licensedProducts: ['jira'],
         adminLinks: ['discoverMore'],
-        fixedLinks: ['people'],
+        fixedLinks: [],
         numberOfSites: 1,
       },
     });
@@ -185,9 +178,9 @@ describe('Atlassian Switcher - Component Analytics', () => {
         data: {
           itemType: 'product',
           itemId: 'jira',
-          itemsCount: 6,
+          itemsCount: 5,
           groupItemIndex: 0,
-          groupItemsCount: 4,
+          groupItemsCount: 3,
           domain: 'invalid',
         },
       },
@@ -196,9 +189,9 @@ describe('Atlassian Switcher - Component Analytics', () => {
         data: {
           itemType: 'product',
           itemId: 'jira',
-          itemsCount: 6,
+          itemsCount: 5,
           groupItemIndex: 0,
-          groupItemsCount: 4,
+          groupItemsCount: 3,
           domain: 'invalid',
         },
         subject: 'atlassianSwitcherItemExpand',
@@ -208,20 +201,9 @@ describe('Atlassian Switcher - Component Analytics', () => {
         data: {
           itemType: 'try',
           itemId: 'confluence.ondemand',
-          itemsCount: 6,
+          itemsCount: 5,
           groupItemIndex: 1,
-          groupItemsCount: 4,
-          domain: 'invalid',
-        },
-      },
-      {
-        name: 'for fixedLinks',
-        data: {
-          itemType: 'product',
-          itemId: 'people',
-          itemsCount: 6,
-          groupItemIndex: 2,
-          groupItemsCount: 4,
+          groupItemsCount: 3,
           domain: 'invalid',
         },
       },
@@ -230,9 +212,9 @@ describe('Atlassian Switcher - Component Analytics', () => {
         data: {
           itemType: 'admin',
           itemId: 'discoverMore',
-          itemsCount: 6,
-          groupItemIndex: 3,
-          groupItemsCount: 4,
+          itemsCount: 5,
+          groupItemIndex: 2,
+          groupItemsCount: 3,
           domain: 'invalid',
         },
       },
@@ -242,7 +224,7 @@ describe('Atlassian Switcher - Component Analytics', () => {
           group: 'recent',
           itemType: 'recent',
           itemId: 'container-type',
-          itemsCount: 6,
+          itemsCount: 5,
           groupItemIndex: 0,
           groupItemsCount: 1,
           domain: 'invalid',
@@ -254,7 +236,7 @@ describe('Atlassian Switcher - Component Analytics', () => {
           group: 'customLinks',
           itemType: 'customLink',
           itemId: null,
-          itemsCount: 6,
+          itemsCount: 5,
           groupItemIndex: 0,
           groupItemsCount: 1,
           domain: 'example.com',
@@ -288,9 +270,9 @@ describe('Atlassian Switcher - Component Analytics', () => {
     const analyticsData = {
       itemType: 'product',
       itemId: 'jira',
-      itemsCount: 6,
+      itemsCount: 5,
       groupItemIndex: 0,
-      groupItemsCount: 4,
+      groupItemsCount: 3,
       domain: 'invalid',
       productType: WorklensProductType.JIRA_BUSINESS,
     };
@@ -355,7 +337,7 @@ describe('Atlassian Switcher - Component Analytics', () => {
       actionSubject: 'button',
     });
     expect(flattenContext(context)).toMatchObject({
-      itemsCount: 6,
+      itemsCount: 5,
     });
   });
 });